move color swatches
