+++ src/packages/property-editors/color-swatches-editor/manifests.ts
