+++ src/packages/property-editors/color-swatches-editor/property-editor-ui-color-swatches-editor.element.ts
